poprawkiw zwiazane ze zmiana API w DAO (Integer -> Long, wiec trzeba bylo dodac .intValue() w kilku miejscach)

+++ src/main/java/de/berlios/jhelpdesk/web/ticket/TicketsNotifiedByMeViewController.java
@@ -144,7 +144,7 @@ public class TicketsNotifiedByMeViewController extends SimpleFormController {
                     ff.setSaviours(filterForm.getSaviours());
                     ff.setStatuses(filterForm.getStatuses());
                     PagingParamsEncoder enc = new PagingParamsEncoder("ticketsIterator", "p_id", request, PAGE_SIZE);
-                    refData.put("ticketsListSize", ticketDao.countTicketsWithFilter(ff));
+                    refData.put("ticketsListSize", ticketDao.countTicketsWithFilter(ff).intValue());
                     refData.put("tickets", ticketDao.getTicketsWithFilter(ff, PAGE_SIZE, enc.getOffset()));
                 } else {
                     refData.put("tickets", ticketDao.getAllTickets());
@@ -189,7 +189,7 @@ public class TicketsNotifiedByMeViewController extends SimpleFormController {
                 // ff.setSaviours( filterForm.getSaviours() );
                 ff.setStatuses(filterForm.getStatuses());
                 PagingParamsEncoder enc = new PagingParamsEncoder("ticketsIterator", "p_id", request, PAGE_SIZE);
-                refData.put("ticketsListSize", ticketDao.countTicketsWithFilter(ff));
+                refData.put("ticketsListSize", ticketDao.countTicketsWithFilter(ff).intValue());
                 refData.put("tickets", ticketDao.getTicketsWithFilter(ff, PAGE_SIZE, enc.getOffset()));
             } else {
                 refData.put("tickets", ticketDao.getAllTickets());
